feat: Admin user sessions!
**Note:** Not stable and not finished yet!

+++ image-api/src/model/role.rs
@@ -2,10 +2,10 @@ use serde::{ Serialize, Deserialize };
 
 use crate::model::user_permissions::UserPermissions;
 
-#[derive(Serialize, Deserialize)]
+#[derive(Default, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Role {
-    pub id: u16,
+    pub id: u8,
     pub role_name: String,
     pub permissions: UserPermissions,
 }
